+++ release/cjs/index.cjs.d.ts
@@ -1,2 +1,3 @@
-import { path } from './path.js';
+import { path } from './core/types.js';
+
 export = path;