+++ release/cjs/index.cjs.js
@@ -1,6 +1,3 @@
 "use strict";
-const path_js_1 = require("./path.js");
-module.exports = path_js_1.path;
-// As {release/package.json}.types = "./esm/index.esm.d.ts",
-// there is no need to export types here, just export values.
+module.exports = require('./path.js').path;
 //# sourceMappingURL=index.cjs.js.map