+++ release/cjs/win32.cjs.d.ts
@@ -1,2 +1,3 @@
-import { pathWin32 } from './path.js';
-export = pathWin32;
+import { path } from './core/types.js';
+
+export = path;