+++ release/cjs/win32.cjs.js
@@ -1,4 +1,3 @@
 "use strict";
-const path_js_1 = require("./path.js");
-module.exports = path_js_1.pathWin32;
+module.exports = require('./path.js').pathWin32;
 //# sourceMappingURL=win32.cjs.js.map