+++ release/esm/path.d.ts
@@ -1,4 +1,3 @@
-export declare const path: typeof import("./core/types.js").path;
-export declare const pathPosix: typeof import("./core/types.js").path;
-export declare const pathWin32: typeof import("./core/types.js").path;
-export type { PathNice, PathNiceArr, ParsedPathNice, FileSystem, PlatformPath, ParsedPath, FormatInputPathObject, } from './core/types.js';
+export declare const path: typeof import("./index.esm.js").default;
+export declare const pathPosix: typeof import("./index.esm.js").default;
+export declare const pathWin32: typeof import("./index.esm.js").default;