+++ release/esm/win32.esm.js
@@ -1,7 +1,7 @@
-import { pathWin32 } from './path.js';
-export default pathWin32;
+import { pathWin32 as path } from './path.js';
+export default path;
 // prettier-ignore
-const { delimiter, sep, posix, win32, PathNice, PathNiceArr, basename, dirname, extname, format, isAbsolute, join, normalize, parse, relative, resolve, toNamespacedPath, } = pathWin32;
+const { bindFS, posix, win32, PathNice, PathNiceArr, delimiter, sep, basename, dirname, extname, format, isAbsolute, join, normalize, parse, relative, resolve, toNamespacedPath } = path;
 // prettier-ignore
-export { delimiter, sep, posix, win32, basename, dirname, extname, format, isAbsolute, join, normalize, parse, relative, resolve, toNamespacedPath, };
+export { bindFS, posix, win32, PathNice, PathNiceArr, delimiter, sep, basename, dirname, extname, format, isAbsolute, join, normalize, parse, relative, resolve, toNamespacedPath };
 //# sourceMappingURL=win32.esm.js.map